# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Dam Sen Smart Guide API
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
...
  /v1/admin/pois:
    get:
      operationId: listAdminPois
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: All POIs including unpublished records
          content:
            application/json:
              schema:
                {
                  type: array,
                  items: { $ref: '#/components/schemas/AdminPoi' },
                }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createAdminPoi
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminPoiInput' }
      responses:
        '201':
          description: Draft created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPoi' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/admin/pois/{id}:
    patch:
      operationId: updateAdminPoi
      security: [{ bearerAuth: [] }]
      parameters:
        [
          {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          },
        ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminPoiInput' }
      responses:
        '200':
          description: Draft updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPoi' }
    delete:
      operationId: deleteAdminPoi
      security: [{ bearerAuth: [] }]
      parameters:
        [
          {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          },
        ]
      responses:
        '204': { description: Draft or rejected POI deleted }
  /v1/admin/pois/{id}/submit:
    post:
      operationId: submitAdminPoi
      security: [{ bearerAuth: [] }]
      parameters:
        [
          {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          },
        ]
      responses:
        '200':
          description: Content version pending review
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPoi' }
  /v1/admin/content/{versionId}/approve:
    post:
      operationId: approveContent
      security: [{ bearerAuth: [] }]
      parameters:
        [
          {
            name: versionId,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          },
        ]
      responses:
        '200':
          description: Version published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPoi' }
  /v1/admin/content/{versionId}/reject:
    post:
      operationId: rejectContent
      security: [{ bearerAuth: [] }]
      parameters:
        [
          {
            name: versionId,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          },
        ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkflowReasonRequest' }
      responses:
        '200':
          description: Version rejected with reason
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPoi' }
  /v1/admin/audit-logs:
    get:
      operationId: listAuditLogs
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: POI mutation audit trail; ADMIN only
          content:
            application/json:
              schema:
                {
                  type: array,
                  items: { $ref: '#/components/schemas/AuditLogEntry' },
                }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/pois:
    get:
      operationId: listPois
      summary: List published POIs
      parameters:
        - {
            name: lat,
            in: query,
            schema: { type: number, minimum: -90, maximum: 90 },
          }
        - {
            name: lng,
            in: query,
            schema: { type: number, minimum: -180, maximum: 180 },
          }
        - {
            name: radius,
            in: query,
            schema: { type: number, minimum: 1, maximum: 5000 },
          }
        - {
            name: category,
            in: query,
            schema: { type: string, pattern: '^[a-z][a-z0-9_-]{0,49}$' },
          }
        - { name: openNow, in: query, schema: { type: boolean } }
        - {
            name: locale,
            in: query,
            schema: { $ref: '#/components/schemas/SupportedLocale' },
          }
      responses:
        '200':
          description: Published POIs, ordered by distance when coordinates are present
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PoiListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /v1/pois/{id}:
    get:
      operationId: getPoi
      summary: Get one published POI
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: locale
          in: query
          schema: { $ref: '#/components/schemas/SupportedLocale' }
      responses:
        '200':
          description: Published POI detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PoiDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404':
          description: POI does not exist or is not published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
...
  /v1/admin/pois/{poiId}/narrations:
    get:
      operationId: listAdminNarrations
      security: [{ bearerAuth: [] }]
      parameters:
        - name: poiId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: All narration revisions, including unpublished revisions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminNarration' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createAdminNarration
      security: [{ bearerAuth: [] }]
      parameters:
        - name: poiId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NarrationInput' }
      responses:
        '201':
          description: Draft narration revision created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminNarration' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/admin/media/presign:
    post:
      operationId: createAudioUploadIntent
      summary: Create a five-minute signed PUT for one immutable audio object
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MediaUploadIntentRequest' }
      responses:
        '201':
          description: Policy-bound upload intent; all required headers must be sent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaUploadIntent' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503':
          description: Object storage is disabled or unavailable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
  /v1/admin/narrations/{id}:
    patch:
      operationId: updateAdminNarration
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NarrationUpdate' }
      responses:
        '200':
          description: Draft or rejected narration updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminNarration' }
    delete:
      operationId: deleteAdminNarration
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Draft or rejected narration deleted }
  /v1/admin/narrations/{id}/submit:
    post:
      operationId: submitAdminNarration
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Narration submitted for review
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminNarration' }
  /v1/admin/narrations/{id}/approve:
    post:
      operationId: approveAdminNarration
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Pending narration published; previous revision superseded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminNarration' }
  /v1/admin/narrations/{id}/reject:
    post:
      operationId: rejectAdminNarration
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkflowReasonRequest' }
      responses:
        '200':
          description: Pending narration rejected with a reason
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminNarration' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Query validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
...
    Forbidden:
      description: Authenticated user lacks the required role
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
...
  schemas:
...
    AdminPoiInput:
      type: object
      required:
        [slug, category, location, translations, entrances, operatingHours]
      properties:
        slug: { type: string }
        category: { type: string }
        location: { $ref: '#/components/schemas/GeoPoint' }
        translations: { type: array, items: { type: object } }
        entrances: { type: array, items: { type: object } }
        operatingHours:
          {
            type: array,
            items: { $ref: '#/components/schemas/PoiOperatingHours' },
          }
    AdminPoi:
      allOf:
        - { $ref: '#/components/schemas/AdminPoiInput' }
        - type: object
          required: [id, status]
          properties:
            id: { type: string, format: uuid }
            status:
              {
                type: string,
                enum: [draft, pending_review, published, rejected],
              }
            pendingVersionId: { type: string, format: uuid }
            rejectionReason: { type: string }
    WorkflowReasonRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 3, maxLength: 1000 }
    AuditLogEntry:
      type: object
      required:
        [id, actorId, action, entityType, entityId, before, after, createdAt]
      properties:
        id: { type: string, format: uuid }
        actorId: { type: string, format: uuid }
        action: { type: string }
        entityType: { type: string, enum: [poi] }
        entityId: { type: string, format: uuid }
        before: { nullable: true }
        after: { nullable: true }
        createdAt: { type: string, format: date-time }
    SupportedLocale:
      type: string
      enum: [vi, en]
      default: vi
...
    GeoPoint:
      type: object
      required: [latitude, longitude]
      properties:
        latitude: { type: number }
        longitude: { type: number }
...
    NarrationAudioInput:
      type: object
      required:
        [
          objectKey,
          mimeType,
          sizeBytes,
          sha256,
          durationSeconds,
          rightsOwner,
          rightsSource,
          usageRights,
        ]
      properties:
        objectKey:
          type: string
          description: Immutable private object key; never returned by the public API
        mimeType:
          type: string
          enum: [audio/mpeg, audio/mp4, audio/ogg, audio/wav]
        sizeBytes: { type: integer, minimum: 1, maximum: 52428800 }
        sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
        durationSeconds: { type: number, exclusiveMinimum: 0, maximum: 1800 }
        rightsOwner: { type: string }
        rightsSource: { type: string }
        usageRights: { type: string }
    MediaUploadIntentRequest:
      type: object
      required: [poiId, locale, mimeType, sizeBytes, sha256]
      properties:
        poiId: { type: string, format: uuid }
        locale: { $ref: '#/components/schemas/SupportedLocale' }
        mimeType:
          type: string
          enum: [audio/mpeg, audio/mp4, audio/ogg, audio/wav]
        sizeBytes: { type: integer, minimum: 1, maximum: 52428800 }
        sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
    MediaUploadIntent:
      type: object
      required: [objectKey, uploadUrl, method, expiresAt, requiredHeaders]
      properties:
        objectKey: { type: string }
        uploadUrl: { type: string, format: uri }
        method: { type: string, enum: [PUT] }
        expiresAt: { type: string, format: date-time }
        requiredHeaders:
          type: object
          required:
            [
              content-type,
              content-length,
              x-amz-checksum-sha256,
              x-amz-meta-sha256,
            ]
          properties:
            content-type: { type: string }
            content-length: { type: string }
            x-amz-checksum-sha256: { type: string }
            x-amz-meta-sha256: { type: string }
    NarrationInput:
      type: object
      required: [locale, transcript]
      properties:
        locale: { $ref: '#/components/schemas/SupportedLocale' }
        transcript: { type: string, minLength: 20, maxLength: 20000 }
        audio:
          allOf: [{ $ref: '#/components/schemas/NarrationAudioInput' }]
          nullable: true
    NarrationUpdate:
      type: object
      properties:
        transcript: { type: string, minLength: 20, maxLength: 20000 }
        audio:
          allOf: [{ $ref: '#/components/schemas/NarrationAudioInput' }]
          nullable: true
    AdminNarration:
      allOf:
        - { $ref: '#/components/schemas/NarrationInput' }
        - type: object
          required: [id, poiId, revision, status, createdAt, updatedAt]
          properties:
            id: { type: string, format: uuid }
            poiId: { type: string, format: uuid }
            revision: { type: integer, minimum: 1 }
            status:
              type: string
              enum: [draft, pending_review, published, rejected, superseded]
            rejectionReason: { type: string }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...
    PoiSummary:
      type: object
      required:
        [
          id,
          slug,
          category,
          location,
          requestedLocale,
          resolvedLocale,
          fallbackUsed,
          name,
          shortDescription,
          isOpen,
        ]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        category: { type: string }
        location: { $ref: '#/components/schemas/GeoPoint' }
        requestedLocale: { $ref: '#/components/schemas/SupportedLocale' }
        resolvedLocale: { $ref: '#/components/schemas/SupportedLocale' }
        fallbackUsed: { type: boolean }
        name: { type: string }
        shortDescription: { type: string }
        distanceMeters: { type: integer, minimum: 0 }
        isOpen: { type: boolean }
...
    PoiEntrance:
      type: object
      required: [id, label, location, graphNodeRef, isPrimary, accessibility]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        location: { $ref: '#/components/schemas/GeoPoint' }
        graphNodeRef: { type: string }
        isPrimary: { type: boolean }
        accessibility: { type: string, enum: [standard, step_free] }
    PoiOperatingHours:
      type: object
      required: [dayOfWeek, opensAt, closesAt]
      properties:
        dayOfWeek: { type: integer, minimum: 0, maximum: 6 }
        opensAt: { type: string }
        closesAt: { type: string }
    PoiDetail:
      allOf:
        - { $ref: '#/components/schemas/PoiSummary' }
        - type: object
          required: [longDescription, entrances, operatingHours]
          properties:
            longDescription: { type: string }
            entrances:
              type: array
              items: { $ref: '#/components/schemas/PoiEntrance' }
            operatingHours:
              type: array
              items: { $ref: '#/components/schemas/PoiOperatingHours' }
    PoiListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/PoiSummary' }
        total: { type: integer, minimum: 0 }
...
    ApiErrorEnvelope:
      type: object
      required: [code, message, details, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        details: { nullable: true }
        requestId: { type: string }